feat: добавить обработку ошибок

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -3,9 +3,11 @@ package ru.yandex.practicum.filmorate.controller;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
 import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.FilmService;
+import ru.yandex.practicum.filmorate.service.UserService;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -15,10 +17,12 @@ import java.util.List;
 @Slf4j
 public class FilmController {
     private final FilmService service;
+    private final UserService userService;
 
     @Autowired
-    public FilmController(FilmService service) {
+    public FilmController(FilmService service, UserService userService) {
         this.service = service;
+        this.userService = userService;
     }
 
     @PostMapping
@@ -41,19 +45,21 @@ public class FilmController {
         return service.getFilm(id);
     }
 
-    @PutMapping("/films/{id}/like/{userId}")
+    @PutMapping("/{id}/like/{userId}")
     public Film addLike(@PathVariable int id, @PathVariable int userId) {
+        if (userService.getUser(userId) == null) throw new NotFoundException("Пользователь не найден");
         return service.addLike(id, userId);
     }
 
-    @DeleteMapping("/films/{id}/like/{userId}")
+    @DeleteMapping("/{id}/like/{userId}")
     public Film removeLike(@PathVariable int id, @PathVariable int userId) {
+        if (userService.getUser(userId) == null) throw new NotFoundException("Пользователь не найден");
         return service.removeLike(id, userId);
     }
 
-    @GetMapping("/films/popular")
+    @GetMapping("/popular")
     public List<Film> getPopular(@RequestParam(required = false) String count) {
-        return service.getTopFilms(Integer.valueOf(count));
+        return service.getTopFilms(count);
     }
 
 }
